+++ src/Admin/AdminUserPagesModel.php
@@ -14,6 +14,9 @@ use Webmozart\Assert\Assert;
 
 /**
  * Count active pages by user.
+ * Route pt/user-pages
+ *
+ * TODO figure out whether part of it may be in Seablast or it is always app dependant
  */
 class AdminUserPagesModel implements SeablastModelInterface
 {